improve code structure and do umentation
